Better module structure, and command line access.
